requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
certifi==2025.1.31
click==8.1.8
coverage==7.8.0
dill==0.4.0
dnspython==2.7.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
motor==3.7.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
platformdirs==4.3.7
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
pylint==3.3.6
pymongo==4.12.0
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
